refactor: rename some files to be more conventional
