+++ frontend/app/components/carousel/carousel-hero-dot-button.tsx
